requirements.txt
pytesseract
pandas
pillow
opencv-python-headless
telethon